maestro-hive/maestro_ml/api/requirements.txt
# Maestro ML Platform API Requirements

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML and MLflow
mlflow==2.10.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
aiohttp==3.9.1
httpx==0.26.0

# Data validation
email-validator==2.1.0

# Environment
python-dotenv==1.0.0
